+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Fixes
+
+* Fix *ConnPool.Deallocate() not deleting prepared statement from map
+
 # 2.8.1 (March 24, 2016)
 
 ## Features